pyproject.toml
[tool.poetry]
name = "octopus-energy"
version = "0.1.2"
description = "Python client for the Octopus Energy RESTful API"
authors = ["Mark Allanson <mark@allanson.org>"]
license = "MIT"
readme = "readme.md"
homepage="https://github.com/markallanson/octopus-energy"
include=["*.md"]

[tool.poetry.dependencies]
python = "^3.7"
python-dateutil = "^2.8.1"
aiohttp = "^3.7.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
requests-mock = "^1.8.0"
black = "^20.8b1"
flake8 = "^3.8.4"
tox = "^3.20.1"
aioresponses = "^0.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py37']
